Cleaning & testing for new NFL weather data
Create a short test that checks if the game is completed. If so, extract those rows, convert to a data frame, and prepare to append the object to our database.
Create a unique event ID column composed of 'date-home-away'. This may need to be changed to 'year-week-home-away' since the date disappears from the time column once the game is completed. The unique event ID will be used to make sure we are not appending duplicate rows to our database.